requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
jinja2
python-multipart
bcrypt
python-jose[cryptography]
itsdangerous
aiofiles
psycopg2-binary
